+++ Common/BusinessLayer/Department.vb
@@ -21,7 +21,7 @@ Namespace BusinessLayer
         Public Property ParentIdNo() As Integer?
         Public Property DepartmentNameAra() As String
         Public Property Notes() As String
-        Public Property RevCostCenterIdNo() As Integer
+        Public Property RevCostCenterIdNo() As Int16
         Public Property SortKey() As String
     End Class
